Fix #61
The fix of #57 makes an assumption that all items contain a
description tag. But it is not always the case and for those item
`map(res_entry, "description", .default = def)` will be `NA` (the
`def`) and the equality operator in the subsequent predicate function dies.

I've included a simple test.